shrink the controller
